separated ajax as a component
